[IMP] point_of_sale: improve topbar
In this commit we improve the topbar from the `product_screen` in pos
(the one containing the category selector and the search input ) such
that it automatically hides when the user is scrolling down and reappears
when the user is scrolling up.
This makes the pos app more fluent to use and provides the benefit
of allowing the user to see more products on the screen while scrolling
down.

This change is needed because the height of the topbar becomes excessive
when we use product images. Hiding the topbar when scrolling solves this issue.

Additionally, we refactor some of the code in order to simplify it.
Thus, the `productList` component is removed and it's contents are put
in the `productScreen` component.

We also remove dead code.

Task: 3595007
