--- addons/point_of_sale/static/src/app/screens/product_screen/product_list/product_list.scss
@@ -1,13 +1,0 @@
-.product-list {
-    grid-template-columns: repeat(auto-fill, minmax(122px, 1fr));
-}
-
-@media only screen and (max-width: 768px) {
-    .product-list.grid {
-        grid-template-columns: repeat(3, 1fr);
-    }
-}
-
-.button-no-demo {
-    color: #017e84 !important;
-}
