+++ addons/pos_restaurant/static/src/overrides/components/product_screen/product_screen.xml
@@ -15,7 +15,7 @@
             <t t-else="">
                 <button
                     class="btn-switchpane btn flex-fill rounded-0 fw-bolder"
-                    t-attf-class="{{ primaryPayButton ? 'btn-primary' : 'btn-secondary' }}"
+                    t-attf-class="{{ currentOrder.is_empty() ? 'btn-secondary' : 'btn-primary' }}"
                     t-on-click="() => currentOrder.pay()">
                     <span class="fs-1 d-block">Pay</span>
                     <span><t t-esc="total" /></span>
