--- addons/pos_self_order/static/src/app/pages/product_list_page/product_list_page.scss
@@ -1,20 +1,0 @@
-.navbar-container {
-    .category-item, .search-button {
-        min-width: calc(3.5rem + 2vw);
-    }
-
-    .category-item {
-        img {
-            background-color: #f7f7f7;
-        }
-
-        &.active {
-            color: $headings-color;
-
-            div {
-                outline: 1px solid $o-action;
-                border-radius: 0.25rem;
-            }
-        }
-    }
-}
